fix(diarization): keep resampled waveform float32 — pyannote died on >2h streams
_wav_to_tensor used np.interp for non-16kHz inputs, which silently returns
float64. Real-world wavs (YouTube 44.1kHz) produced a DoubleTensor; pyannote's
SincNet (wav_norm1d -> cudnn_batch_norm) requires float32, so diarization fell
back to the silence-gap heuristic on every long real recording while synthetic
16k test audio masked the bug.

Fix: cast back to float32 after interpolation + regression test with a real
44.1kHz wav. Verified live: full 2:11 stream now diarizes (4 speakers).

+++ src/meeting_intelligence/transcribe.py
@@ -115,7 +115,16 @@ def _wav_to_tensor(path: Path):
         data = data.reshape(-1, ch).mean(axis=1)
     if sr != 16000:
         n_out = int(round(data.shape[0] * 16000.0 / sr))
-        data = np.interp(np.linspace(0, data.shape[0] - 1, n_out), np.arange(data.shape[0]), data)
+        # np.interp всегда возвращает float64 — pyannote требует float32
+        # (иначе DoubleTensor падает на cudnn_batch_norm). См.
+        # tests/unit/test_wav_to_tensor.py::test_resampled_wav_stays_float32
+        data = (
+            np.interp(
+                np.linspace(0, data.shape[0] - 1, n_out),
+                np.arange(data.shape[0]),
+                data,
+            )
+        ).astype(np.float32)
         sr = 16000
     return torch.from_numpy(data).unsqueeze(0), sr
